perf: 将speaker的check放入父类SpeakerInterface中，bv2使用babala模型
