+++ src/module/speaker/gsv.py
@@ -55,19 +55,3 @@ class GsvSpeaker(SpeakerInterface):
         ]
 
         subprocess.Popen(command, cwd=root_path, shell=True, env=os.environ)
-
-    def check(self) -> Tuple[bool, Exception | None]:
-        total_time = 60
-        iter_time = 10
-        for _ in range(total_time // iter_time):
-            try:
-                output_path = self.speak("你好")
-                if os.path.exists(output_path):
-                    return True, None
-                else:
-                    return False, Exception()
-            except Exception as e:
-                print(e)
-                time.sleep(iter_time)
-
-        return False, TimeoutError()
